添加不同尺寸的CU在INTRA模式上的计时

+++ source/Lib/EncoderLib/EncModeCtrl.h
@@ -294,7 +294,7 @@ protected:
   bool                  m_doPlt;
 
 public:
-  static long ccIntraT;
+  static double ccIntraT;
 
   virtual ~EncModeCtrl              () {}
